+++ CHANGES.md
@@ -1,6 +1,7 @@
 # ChangeLog
 
 ## vx.x.x
+* Scales the displacement vectors keeping the color bar with the displacement values. Adds title to color bar #270
 * Allow inp format in point-cloud file #269 
 * Edit documentation on registration in dvc steps #268
 * Edit documentation on point 0 in dvc steps #264